+++ src/frontend/components/PictureInPictureLayer/PictureInPictureElement/index.tsx
@@ -1,6 +1,7 @@
 import { Box, BoxProps, Stack } from 'grommet';
 import React, {
   CSSProperties,
+  Fragment,
   MutableRefObject,
   ReactNode,
   useState,
@@ -9,6 +10,8 @@ import styled from 'styled-components';
 
 import { startDraggingHandler } from 'components/PictureInPictureLayer/usePIPDragger';
 import { Nullable } from 'utils/types';
+
+import { ResizerCorner } from './ResizerCorner';
 
 const DragLayer = styled(Box)`
   cursor: move;
@@ -21,6 +24,7 @@ interface PictureInPictureElementProps extends BoxProps {
   isPicture?: boolean;
   pictureActions?: ReactNode[];
   startDragging?: startDraggingHandler;
+  startResizing?: startDraggingHandler;
   style?: CSSProperties;
 }
 
@@ -31,6 +35,7 @@ export const PictureInPictureElement = ({
   isPicture,
   pictureActions,
   startDragging,
+  startResizing,
   style,
 }: PictureInPictureElementProps) => {
   const [isHover, setIsHover] = useState(false);
@@ -51,31 +56,35 @@ export const PictureInPictureElement = ({
         fill
       >
         {children}
-        <DragLayer
-          fill
-          onMouseDown={startDragging}
-          hidden={!isPicture}
-          onMouseOver={() => setIsHover(true)}
-          onMouseLeave={() => setIsHover(false)}
-          animation={
-            isHover
-              ? { type: 'fadeIn', duration: 350 }
-              : { type: 'fadeOut', duration: 350, delay: 1800 }
-          }
-          background={
-            pictureActions && pictureActions.length > 0
-              ? '#0000008C'
-              : 'transparent'
-          }
-        >
-          <Box justify="between" direction="row" fill pad="small">
-            {pictureActions?.map((action, index) => (
-              <Box key={`pipi_action_${index}`} margin="auto">
-                {action}
-              </Box>
-            ))}
-          </Box>
-        </DragLayer>
+        <Fragment>
+          <DragLayer
+            fill
+            onMouseDown={startDragging}
+            hidden={!isPicture}
+            onMouseOver={() => setIsHover(true)}
+            onMouseLeave={() => setIsHover(false)}
+            animation={
+              isHover
+                ? { type: 'fadeIn', duration: 350 }
+                : { type: 'fadeOut', duration: 350, delay: 1800 }
+            }
+            background={
+              pictureActions && pictureActions.length > 0
+                ? '#0000008C'
+                : 'transparent'
+            }
+          >
+            <Box justify="between" direction="row" fill pad="small">
+              {pictureActions?.map((action, index) => (
+                <Box key={`pipi_action_${index}`} margin="auto">
+                  {action}
+                </Box>
+              ))}
+            </Box>
+          </DragLayer>
+
+          {isPicture && <ResizerCorner startResizing={startResizing} />}
+        </Fragment>
       </Stack>
     </Box>
   );