+++ CHANGELOG.md
@@ -18,6 +18,7 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 - stopped_at is deleted and started_at is reinitialized when the
 live is restarted
 - Display number of anonymous instead of their pseudos
+- Rename markdown route in markdown-documents
 
 
 ## [4.0.0-beta.6] - 2022-06-20